chore: upgrade ts 5.0 tc
