+++ package.json
@@ -86,7 +86,7 @@
     "prettier": "^2.8.4",
     "svgo": "^3.0.2",
     "ts-node": "^10.9.1",
-    "typescript": "5.0.0-dev.20221202",
+    "typescript": "5.1.0-dev.20230302",
     "vite": "^4.1.4",
     "vitest": "^0.29.1"
   },
@@ -101,7 +101,8 @@
       "@mui/base": "5.0.0-alpha.100",
       "@protobufjs/inquire": "1.1.0",
       "@ethersproject/providers": "5.7.2",
-      "reflect-metadata": "0.1.13"
+      "reflect-metadata": "0.1.13",
+      "@storybook/react-docgen-typescript-plugin": "npm:react-docgen-typescript-plugin@1.0.2"
     },
     "override comments": {
       "ipfs-utils": "see https://github.com/ipfs/js-ipfs-utils/issues/158",
