+++ packages/config/tsconfig.json
@@ -1,5 +1,6 @@
 {
   "compilerOptions": {
+    // "verbatimModuleSyntax": true,
     // Classification follows https://www.typescriptlang.org/tsconfig
 
     // Type Checking
@@ -23,9 +24,6 @@
     // Emit
     "declaration": true,
     "declarationMap": true,
-    "importsNotUsedAsValues": "error",
-    "noEmitOnError": true,
-    "preserveValueImports": true,
     "sourceMap": true,
     "stripInternal": true, // skip type emit for @internal types
