+++ packages/mask/src/plugins/Savings/SNSAdaptor/SavingsForm.tsx
@@ -2,7 +2,7 @@ import { useState, useMemo } from 'react'
 import { useAsync, useAsyncFn } from 'react-use'
 import type { AbiItem } from 'web3-utils'
 import { BigNumber } from 'bignumber.js'
-import { isLessThan, rightShift, ZERO, formatBalance, formatCurrency } from '@masknet/web3-shared-base'
+import { isLessThan, rightShift, ZERO, formatBalance, formatCurrency, isPositive } from '@masknet/web3-shared-base'
 import { LoadingBase, makeStyles } from '@masknet/theme'
 import {
     createContract,
@@ -318,7 +318,7 @@ export function SavingsFormDialog({ chainId, protocol, tab, onClose }: SavingsFo
                             ) : (
                                 <Typography variant="body2" textAlign="right" className={classes.tokenValueUSD}>
                                     &asymp; <FormattedCurrency value={tokenValueUSD} formatter={formatCurrency} />
-                                    {estimatedGas > 0 ? (
+                                    {isPositive(estimatedGas) ? (
                                         <span className={classes.gasFee}>+ {formatBalance(estimatedGas, 18)} ETH</span>
                                     ) : (
                                         <span />
