+++ packages/plugin-infra/src/types.ts
@@ -847,6 +847,14 @@ export namespace Plugin.SNSAdaptor {
         }
     }
 
+    export interface SettingsTabUIProps {
+        onClose: () => void
+        onOpenPopup: (route?: PopupRoutes, params?: Record<string, any>) => void
+        bindingWallets?: BindingProof[]
+        currentPersona?: ECKeyIdentifier
+        pluginID: PluginID
+    }
+
     export interface SettingTab {
         ID: PluginID
         /**
@@ -860,13 +868,7 @@ export namespace Plugin.SNSAdaptor {
         priority: number
 
         UI?: {
-            TabContent: InjectUI<{
-                onClose: () => void
-                onOpenPopup: (route?: PopupRoutes, params?: Record<string, any>) => void
-                bindingWallets?: BindingProof[]
-                currentPersona?: ECKeyIdentifier
-                pluginID: PluginID
-            }>
+            TabContent: InjectUI<SettingsTabUIProps>
         }
     }
