+++ packages/polyfills/types/__all.d.ts
@@ -3,3 +3,4 @@
 /// <reference path="./intl.d.ts" />
 /// <reference path="./global.d.ts" />
 /// <reference path="./theme-global.d.cts" />
+/// <reference path="./webpack-hmr.d.ts" />
