+++ tsconfig.json
@@ -29,7 +29,6 @@
     // Exclude everything by default. Include them in each project on-demand
     "types": ["masknet__global-types"],
     "resolveJsonModule": true,
-    "preserveValueImports": false,
     "paths": {
       // # Why?
       // TypeScript project reference requires manually declaring all referenced projects.
